+++ app/lib/service/csp/default_csp.dart
@@ -71,6 +71,7 @@ final _defaultContentSecurityPolicyMap = <String, List<String>>{
   'base-uri': <String>["'self'"],
   'form-action': <String>["'self'"],
   'upgrade-insecure-requests': <String>[],
+  'report-to': <String>['csp-endpoint'],
 };
 
 /// Returns the serialized string of the CSP header.